[Feat] UI 구현

+++ Assets/Scripts/Player/PlayerBase/PlayerStats.cs
@@ -7,7 +7,7 @@ public class PlayerStats
 {
     public ObseravableProperty<int> CurHp = new();
     public ObseravableProperty<int> MaxHp = new();
-    public ObseravableProperty<int> CurHookCount = new ObseravableProperty<int>();
+    public ObseravableProperty<float> CurHookCount = new ObseravableProperty<float>();
     public ObseravableProperty<int> CurBulletCount = new ObseravableProperty<int>();
     // 플레이어 사망 후 이루어지는 것들은 Start에서 구독해야됨. Awake에서 false로 초기화.
     public ObseravableProperty<bool> IsDied = new ObseravableProperty<bool>();
